utils.parser: extracted from filterExpression
