+++ src/utils/filterExpression/lexems.rim
@@ -77,3 +77,26 @@ filterExpressionLanguage is Syntax: expr
 	id: /^[^ .(){}[\]\n\t"]+/
 		strip: id | (id?, (dot, (id | text | paren))*) // abc, .a."b".("b"+c)
 		special: /^[-<>=+*/!,]+/
+
+
+' Notes
+	§1234567890+´¨'-.,<
+	°!"#€%&/()=?`^*_:;>
+	¶©@£$∞§|[]≈±´~™–…‚≤
+	•¡”¥¢‰¶\{}≠¿`^’—·„≥
+	•Ω®†ıπ˙ß∂ƒ¸˛√ªﬁøæ÷≈‹›‘’°˜ˆ∏˚◊∫¯˘¬º«»”
+	<>≤≥
+	'"„”`´’‘«»‹›
+	§1234567890
+	±+-/=≈^*~√÷≈≠
+	,‚:;
+	#‰%&|
+	!?¡¿
+	¶@§–…
+	¶—•·
+	©™®†ı¸˛
+	°˜ˆ˘˚˙¨¯
+	∞Ω¥¢£$€
+	πß∂ƒªﬁøæ
+	∏◊∫¬º
+	()[]{}\\
