+++ src/utils/parser/lexemUtils.js
@@ -117,29 +117,3 @@ export const expand = root=> {
 	// intermediate lexems = named through recursivelyAddNameToLexems
 	// all lexems = state.named + intermediate lexems
 }
-
-
-/* Notes
-§1234567890+´¨'-.,<
-°!"#€%&/()=?`^*_:;>
-¶©@£$∞§|[]≈±´~™–…‚≤
-•¡”¥¢‰¶\{}≠¿`^’—·„≥
-•Ω®†ıπ˙ß∂ƒ¸˛√ªﬁøæ÷≈‹›‘’°˜ˆ∏˚◊∫¯˘¬º«»”
-
-<>≤≥
-'"„”`´’‘«»‹›
-§1234567890
-±+-/=≈^*~√÷≈≠
-,‚:;
-#‰%&|
-!?¡¿
-¶@§–…
-¶—•·
-©™®†ı¸˛
-°˜ˆ˘˚˙¨¯
-∞Ω¥¢£$€
-πß∂ƒªﬁøæ
-∏◊∫¬º
-
-()[]{}\\
-*/
