+++ src/utils/parser/lexemUtils.test.js
